Operation On The Move
Unit Movement has been implemented! This is a huge commit that adds in the template for a critical part of the game.

The observable change is this: when you click on a tile with a unit on it, you can see all the tiles that unit can move to.

Now we have to implement actually moving there when you click on one of THOSE tiles.